+++ changelog.md
@@ -7,6 +7,7 @@
 - changed bean plant to only show when setting enabled
 - changed trade sequences to only show current step
 - added indicators for vanilla small key/map/compass chests
+- added Adult Fountain setting
 
 ## 2.6.0.3